pyproject.toml
[tool.poetry]
name = "ngrams"
version = "0.1.0"
description = ""
authors = ["dev <dev@togglecorp.com>"]

[tool.poetry.dependencies]
python = ">=3.8,<3.9.7 || >3.9.7"
openai = ">=0.27.8"
python-dotenv = "==1.0.0"
streamlit = "==1.23.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
